fix: readme index route
