+++ apps/website/src/app/docs/packages/page.tsx
@@ -5,24 +5,7 @@ import { VscPackage } from '@react-icons/all-files/vsc/VscPackage';
 import Link from 'next/link';
 import { PACKAGES } from '~/util/constants';
 
-async function getData() {
-	return Promise.all(
-		PACKAGES.map(async (pkg) => {
-			const response = await fetch(`https://docs.discordjs.dev/api/info?package=${pkg}`, {
-				next: { revalidate: 3_600 },
-			});
-			const versions = await response.json();
-			const latestVersion = versions.at(-2) ?? 'main';
-			return { packageName: pkg, version: latestVersion };
-		}),
-	);
-}
-
 export default async function Page() {
-	const data = await getData();
-
-	const findLatestVersion = (pkg: string) => data.find((version) => version.packageName === pkg);
-
 	return (
 		<div className="min-w-xs sm:w-md mx-auto flex min-h-screen flex-row place-content-center place-items-center gap-8 py-0 px-4 lg:py-0 lg:px-6">
 			<div className="flex grow flex-col place-content-center gap-4">
@@ -44,7 +27,7 @@ export default async function Page() {
 				{PACKAGES.map((pkg) => (
 					<Link
 						className="dark:bg-dark-400 dark:border-dark-100 dark:hover:bg-dark-300 dark:active:bg-dark-200 focus:ring-width-2 focus:ring-blurple flex h-11 transform-gpu cursor-pointer select-none appearance-none flex-row place-content-between rounded border border-neutral-300 bg-transparent p-4 text-base font-semibold leading-none text-black outline-0 hover:bg-neutral-100 focus:ring active:translate-y-px active:bg-neutral-200 dark:text-white"
-						href={`/docs/packages/${pkg}/${findLatestVersion(pkg)?.version ?? 'main'}`}
+						href={`/docs/packages/${pkg}/`}
 						key={pkg}
 					>
 						<div className="flex grow flex-row place-content-between place-items-center gap-4">
