+++ apps/website/src/util/constants.ts
@@ -11,6 +11,8 @@ export const PACKAGES = [
 	'ws',
 ];
 
+export const N_RECENT_VERSIONS = 2;
+
 export const DESCRIPTION =
 	"discord.js is a powerful node.js module that allows you to interact with the Discord API very easily. It takes a much more object-oriented approach than most other JS Discord libraries, making your bot's code significantly tidier and easier to comprehend.";
